+++ cli_options/CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+* Support getting option value from an environment variable, through the `:env` option. Used
+only if not provided by the user in the CLI arguments.
+
 * Support grouping options in the docs by section. You can now specify the `:doc_section`
 to any option, and pass a `:sections` option in the `CliOptions.docs/2` function for the
 headers and extended docs of each section.